backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
celery[redis]==5.4.0
python-multipart==0.0.9
celery[redis]
httpx

